refactor: rename project pages and update related widget references

+++ lib/src/layer_presentation/pages/project_pages/project_list_page/widgets/projects_action_popup_menu_button.dart
@@ -3,11 +3,14 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:genesis/src/core/extensions/localized_build_context.dart';
 import 'package:genesis/src/layer_domain/entities/project.dart';
 import 'package:genesis/src/layer_presentation/blocs/projects_bloc/projects_bloc.dart';
-import 'package:genesis/src/layer_presentation/pages/projects_page/widgets/delete_projects_dialog.dart';
+import 'package:genesis/src/layer_presentation/shared_widgets/confirmation_dialog.dart';
 import 'package:genesis/src/theming/palette.dart';
 
 class ProjectsActionPopupMenuButton extends StatelessWidget {
-  const ProjectsActionPopupMenuButton({required this.project, super.key});
+  const ProjectsActionPopupMenuButton({
+    required this.project,
+    super.key,
+  });
 
   final Project project;
 
@@ -22,16 +25,14 @@ class ProjectsActionPopupMenuButton extends StatelessWidget {
         onSelected: (value) {
           final projects = List.generate(1, (_) => project);
           final child = switch (value) {
-            _PopupBtnValue.deleteProject => DeleteProjectsDialog(
-              projects: projects,
+            _PopupBtnValue.deleteProject => ConfirmationDialog(
+              message: context.$.deleteProjectConfirmation(project.name),
               onDelete: () => context.read<ProjectsBloc>().add(ProjectsEvent.deleteProjects(projects)),
             ),
           };
           showDialog<void>(
             context: context,
-            builder: (_) {
-              return child;
-            },
+            builder: (_) => child,
           );
         },
         itemBuilder: (context) {
